10) Similarity Search/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
click==8.2.1
fastapi==0.115.13
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
numpy==2.3.0
portalocker==2.10.1
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
qdrant-client==1.14.3
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
